+++ src/components/EditSections/EditFeesFines/EditFeesFines.js
@@ -19,6 +19,7 @@ const propTypes = {
 
 const EditFeesFines = ({
   accordionId,
+  isCreateMode,
 }) => {
   const intl = useIntl();
   const renderState = useCustomFieldsSection({ sectionId: CUSTOM_FIELDS_SECTION.FEES_FINES });
@@ -45,6 +46,7 @@ const EditFeesFines = ({
         <EditCustomFieldsSection
           accordionId={accordionId}
           sectionId={CUSTOM_FIELDS_SECTION.FEES_FINES}
+          isCreateMode={isCreateMode}
         />
       </Row>
     </Accordion>